Changed array size from int to unsigned long everywhere

+++ src/errors.c
@@ -23,7 +23,7 @@ int check_parenthesis(char *str, char *operators)
 {
     int p = 0;
 
-    for (int i = 0; str[i] != '\0'; i++) {
+    for (unsigned long i = 0; str[i] != '\0'; i++) {
         if (str[i] == operators[OP_OPENP])
             p++;
         if (str[i] == operators[OP_CLOSEP])
