+++ src/infin_sub.c
@@ -10,12 +10,12 @@
 
 char *infin_sub(char *nb1, char *nb2, char *base)
 {
-    int size = na_size(nb2) + 1;
+    unsigned long size = na_size(nb2) + 1;
     char *tmp;
     char *result;
 
     tmp = malloc(sizeof(char) * size);
-    for (int i = 0; i < size; i++)
+    for (unsigned long i = 0; i < size; i++)
         tmp[i] = -1 * nb2[i];
     result = infin_add(nb1, tmp, base);
     free(tmp);
